# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.JSONResponse:
    properties:
      data: {}
      error_id:
        type: string
      message:
        $ref: '#/definitions/common.message'
      status:
        type: integer
    type: object
  common.message:
    properties:
      code:
        type: string
      en:
        type: string
      th:
        type: string
    type: object
  constants.ProductStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - ProductStatusActive
    - ProductStatusInactive
  constants.SupplierStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - SupplierStatusActive
    - SupplierStatusInactive
  dto.CreateProductRequest:
    properties:
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
      supplier_id:
        type: integer
    required:
    - name
    - price
    type: object
  dto.CreateSupplierRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.ProductListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ProductResponse'
        type: array
      meta:
        $ref: '#/definitions/internal_product_interface_http_dto.PaginationMeta'
    type: object
  dto.ProductResponse:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/internal_product_interface_http_dto.UserBriefResponse'
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      status:
        $ref: '#/definitions/constants.ProductStatus'
      stock:
        type: integer
      supplier:
        $ref: '#/definitions/internal_product_interface_http_dto.SupplierResponse'
      supplier_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/internal_product_interface_http_dto.UserBriefResponse'
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - firstname
    - lastname
    - password
    - username
    type: object
  dto.SupplierListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_supplier_interface_http_dto.SupplierResponse'
        type: array
      meta:
        $ref: '#/definitions/internal_supplier_interface_http_dto.PaginationMeta'
    type: object
  dto.UpdateProductRequest:
    properties:
      name:
        type: string
      price:
        type: number
      status:
        type: string
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    - status
    type: object
  dto.UpdateSupplierRequest:
    properties:
      name:
        type: string
      status:
        type: string
    required:
    - name
    - status
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  internal_product_interface_http_dto.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_product_interface_http_dto.SupplierResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  internal_product_interface_http_dto.UserBriefResponse:
    properties:
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      username:
        type: string
    type: object
  internal_supplier_interface_http_dto.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_supplier_interface_http_dto.SupplierResponse:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/internal_supplier_interface_http_dto.UserBriefResponse'
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        $ref: '#/definitions/constants.SupplierStatus'
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/internal_supplier_interface_http_dto.UserBriefResponse'
    type: object
  internal_supplier_interface_http_dto.UserBriefResponse:
    properties:
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      username:
        type: string
    type: object
host: localhost:9392
info:
  contact: {}
  description: Product API with Read/Write DB Separation and Redis Cache-Aside Pattern
  title: Product API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.JSONResponse'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.JSONResponse'
      summary: Register user
      tags:
      - auth
  /products:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by supplier
        in: query
        name: supplier_id
        type: integer
      - default: id
        description: Sort field
        enum:
        - id
        - name
        - price
        - stock
        - created_at
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      security:
      - BearerAuth: []
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      summary: Create product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Set deleted_at — product จะไม่แสดงใน query ปกติ แต่ข้อมูลยังอยู่ใน
        DB
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      summary: Soft delete product
      tags:
      - products
    get:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.JSONResponse'
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      summary: Update product
      tags:
      - products
  /products/{id}/hard:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      security:
      - BearerAuth: []
      summary: Hard delete product
      tags:
      - products
  /suppliers:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: id
        description: Sort field
        enum:
        - id
        - name
        - created_at
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SupplierListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      security:
      - BearerAuth: []
      summary: List suppliers
      tags:
      - suppliers
    post:
      consumes:
      - application/json
      parameters:
      - description: Supplier data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSupplierRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_supplier_interface_http_dto.SupplierResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      security:
      - BearerAuth: []
      summary: Create supplier
      tags:
      - suppliers
  /suppliers/{id}:
    delete:
      consumes:
      - application/json
      description: Set deleted_at — supplier จะไม่แสดงใน query ปกติ แต่ข้อมูลยังอยู่ใน
        DB
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      security:
      - BearerAuth: []
      summary: Soft delete supplier
      tags:
      - suppliers
    get:
      consumes:
      - application/json
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_supplier_interface_http_dto.SupplierResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.JSONResponse'
      security:
      - BearerAuth: []
      summary: Get supplier by ID
      tags:
      - suppliers
    put:
      consumes:
      - application/json
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      - description: Supplier data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSupplierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_supplier_interface_http_dto.SupplierResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      security:
      - BearerAuth: []
      summary: Update supplier
      tags:
      - suppliers
  /suppliers/{id}/hard:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.JSONResponse'
      security:
      - BearerAuth: []
      summary: Hard delete supplier
      tags:
      - suppliers
securityDefinitions:
  BearerAuth:
    description: 'JWT token format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"